\name{badree}
\alias{badree}
\docType{data}
\title{
Data set for Samuel Badree
}
\description{
CSV file Samuel Badree
}
\usage{data("badree")}
\format{
  The format is:
 chr "badree"
}
\details{
CSV file Samuel Badree
}
\source{
ESPN Cricinfo Statsguru
}
\references{
https://www.espncricinfo.com/ci/content/stats/index.html
}
\keyword{datasets}
