% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DL_funs.R
\name{clean.rtab}
\alias{clean.rtab}
\title{clean/reformat resource table
FIXME: don't hard-code max overs (wickets)}
\usage{
clean.rtab(rtab, corner = 100, edge = edge, flip = TRUE)
}
\arguments{
\item{rtab}{resource table}

\item{corner}{}

\item{edge}{}

\item{flip}{}
}
\description{
clean/reformat resource table
FIXME: don't hard-code max overs (wickets)
}

