% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanTeamData.R
\name{cleanTeamData}
\alias{cleanTeamData}
\title{Clean the team data for Test, ODI and T20}
\usage{
cleanTeamData(df,matchType)
}
\arguments{
\item{df}{Data frame}

\item{matchType}{Match type - Test, ODI, T20}
}
\value{
The cleaned Data frame
}
\description{
This function cleans the team data for Test, ODI and T20
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
\dontrun{
#Get the team data for India for Tests
df<-getTeamDataHomeAway(file="india.csv",teamName="India",matchType='Test')
df1 <-cleanTeamData(df,"Test")
}
}
\references{
\url{http://www.espncricinfo.com/ci/content/stats/index.html}\cr
\url{https://gigadom.in/}\cr
}
\seealso{
\code{\link{teamWinLossStatusVsOpposition}}
\code{\link{teamWinLossStatusAtGrounds}}
\code{\link{plotTimelineofWinsLosses}}
}
\author{
Tinniam V Ganesh
}
