% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/DL_funs.R
\name{cross.valid}
\alias{cross.valid}
\title{run a single cross-validation}
\usage{
cross.valid(dat, samp = 0.8)
}
\arguments{
\item{dat}{data set: should include both innings for subset of games}

\item{samp}{sampling fraction}

\item{dat.sub}{}
}
\description{
run a single cross-validation
}

