\name{devilliers}
\alias{devilliers}
\docType{data}
\title{
Data set for AB Devilliers
}
\description{
Data set for AB Devilliers
}
\usage{data("devilliers")}
\format{
  The format is:
 chr "devilliers"
}
\details{
Data set for AB Devilliers
}
\source{
http://www.espncricinfo.com/ci/content/stats/index.html
}
\references{
http://www.espncricinfo.com/ci/content/stats/index.html
}
\keyword{datasets}
