% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DL_funs.R
\name{DLtab}
\alias{DLtab}
\title{Duckworth-Lewis resource table}
\usage{
DLtab(p, tot.runs, max.over = 50)
}
\arguments{
\item{tot.runs}{avg number of runs across all games}

\item{max.over}{max number of overs}
}
\description{
Duckworth-Lewis resource table
}

