% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/DL_funs.R
\name{dnorm2}
\alias{dnorm2}
\title{normal density with profiled sd}
\usage{
dnorm2(x, mean, log = TRUE)
}
\arguments{
\item{x}{data}

\item{mean}{}

\item{log}{return log-density?}
}
\description{
normal density with profiled sd
}

