\name{dravid}
\alias{dravid}
\docType{data}
\title{
Data set for Rahul Dravid
}
\description{
Data set for Rahul Dravid
}
\usage{data("dravid")}
\format{
  The format is:
 chr "dravid"
}
\details{
Data set for Rahul Dravid
}
\source{
http://www.espncricinfo.com/ci/content/stats/index.html
}
\references{
http://www.espncricinfo.com/ci/content/stats/index.html
}
\keyword{datasets}
