% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/DL_funs.R
\name{get.res}
\alias{get.res}
\title{compute resources for a specific set: new hotness (dplyr)}
\usage{
get.res(sub, full, fun = mean, rmult = 100)
}
\arguments{
\item{sub}{subset of data: e.g. single over/wicket combination}

\item{full}{full data}

\item{fun}{summarizing function}

\item{rmult}{scaling factor for table}
}
\description{
FIXME: can we compute finalruns in advance?
}

