% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/DL_funs.R
\name{get.rtab}
\alias{get.rtab}
\title{calculate resource table}
\usage{
get.rtab(dd, fun = mean, corner = 100, edge = NA)
}
\arguments{
\item{dd}{data frame}

\item{fun}{summarizing function}

\item{corner}{value for max wicket/over combination}

\item{edge}{value for impossible combinations}
}
\description{
calculate resource table
}

