% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMatchType.R
\name{getMatchType}
\alias{getMatchType}
\title{Get the number of the match type viz.for Test, ODI and T20}
\usage{
getMatchType(matchType)
}
\arguments{
\item{matchType}{The match type - Test, ODI or T20}
}
\value{
The numerical value of match type
}
\description{
This function returns the number of the match type
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
\dontrun{
#Get the team data for India for Tests

match <-getMatchType("Test")
}
}
\references{
\url{http://www.espncricinfo.com/ci/content/stats/index.html}\cr
\url{https://gigadom.in/}\cr
}
\seealso{
\code{\link{teamWinLossStatusVsOpposition}}
\code{\link{teamWinLossStatusAtGrounds}}
\code{\link{plotTimelineofWinsLosses}}
}
\author{
Tinniam V Ganesh
}
