% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTeamDataHomeAway.R
\name{getTeamDataHomeAway}
\alias{getTeamDataHomeAway}
\title{Get the data for a team in a match type  viz.for Test, ODI and T20 with the home/overseas/neutral}
\usage{
getTeamDataHomeAway(dir=".",teamView="bat",matchType="Test",file="team001HA.csv",
     save=TRUE,teamName)
}
\arguments{
\item{dir}{The directory where the team data CSV file be saved}

\item{teamView}{Team view can be either 'bat' (batting team) or 'bowl' (bowling team)}

\item{matchType}{The match type - Test, ODI , T20}

\item{file}{The name of te file to save to}

\item{save}{This can be TRUE or FALSE}

\item{teamName}{Team name is the team namely - Australia, India, England etc}
}
\value{
The required data frame
}
\description{
This function returns team data as a CSV file and/or a dataframe for Test, ODI and T20 with an additional
column showing home, away or neutral venue where the match was played
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
\dontrun{
#Get the team data for India for Tests

getTeamDataHomeAway(teamName="India",file="india.csv")
}
}
\references{
\url{http://www.espncricinfo.com/ci/content/stats/index.html}\cr
\url{https://gigadom.in/}\cr
}
\seealso{
\code{\link{teamWinLossStatusVsOpposition}}
\code{\link{teamWinLossStatusAtGrounds}}
\code{\link{plotTimelineofWinsLosses}}
}
\author{
Tinniam V Ganesh
}
