% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTeamNumber.R
\name{getTeamNumber}
\alias{getTeamNumber}
\title{Get the number of the Team}
\usage{
getTeamNumber(teamName,matchType)
}
\arguments{
\item{teamName}{The name of the team e.g Australia, India, Ghana etc}

\item{matchType}{The match type - Test, ODI or T20}
}
\value{
The numerical value of the team
}
\description{
This function returns the number of the Team for which analysis is to be done
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
\dontrun{
#Get the team data for India for Tests

teamNi <-getTeamNumber(teamName="India",matchType="Test")
}
}
\references{
\url{https://www.espncricinfo.com/ci/content/stats/index.html}\cr
\url{https://gigadom.in/}\cr
}
\seealso{
\code{\link{teamWinLossStatusVsOpposition}}
\code{\link{teamWinLossStatusAtGrounds}}
\code{\link{plotTimelineofWinsLosses}}
}
\author{
Tinniam V Ganesh
}
