% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/summary_funs.R
\name{gof}
\alias{gof}
\title{Calculate summary statistics (bias, variance, RMSE, kappa) based on a resource table, at a specified stopping point (number of overs).  Uses first inning}
\usage{
gof(stop_overs, rtab_full = res_tab_R_50, max_overs = 50, max_wickets = 9,
  fulldata)
}
\arguments{
\item{stop_overs}{number of overs to check}

\item{rtab_full}{resource table}

\item{max_overs}{max number of overs per game}

\item{fulldata}{full data frame (typically over50.data)}
}
\description{
Calculate summary statistics (bias, variance, RMSE, kappa) based on a resource table, at a specified stopping point (number of overs).  Uses first inning
}

