\name{kohli1}
\alias{kohli1}
\docType{data}
\title{
Data set for Virat Kohli
}
\description{
Data set for Virat Kohli
}
\usage{data("kohli1")}
\format{
  The format is:
 chr "kohli1"
}
\details{
Data set for Virat Kohli
}
\source{
https://www.espncricinfo.com/ci/content/stats/index.html
}
\references{
https://www.espncricinfo.com/ci/content/stats/index.html
}
\keyword{datasets}
