\name{kumble1}
\alias{kumble1}
\docType{data}
\title{
Data set for Anil Kumble
}
\description{
Data set for Anil Kumble
}
\usage{data("kumble1")}
\format{
  The format is:
 chr "kumble1"
}
\details{
Data set for Anil Kumble
}
\source{
https://www.espncricinfo.com/ci/content/stats/index.html
}
\references{
https://www.espncricinfo.com/ci/content/stats/index.html
}
\keyword{datasets}
