\name{malinga}
\alias{malinga}
\docType{data}
\title{
Data set for Lasith Malinga
}
\description{
Data set for Lasith Malinga
}
\usage{data("malinga")}
\format{
  The format is:
 chr "malinga"
}
\details{
Data set for Lasith Malinga
}
\source{
http://www.espncricinfo.com/ci/content/stats/index.html
}
\references{
http://www.espncricinfo.com/ci/content/stats/index.html
}
\examples{
data(malinga)
## maybe str(malinga) ; plot(malinga) ...
}
\keyword{datasets}
