\name{malinga1}
\alias{malinga1}
\docType{data}
\title{
Data set for Lasith Malinga
}
\description{
Data set for Lasith Malinga
}
\usage{data("malinga1")}
\format{
  The format is:
 chr "malinga1"
}
\details{
Data set for Lasith Malinga
}
\source{
https://www.espncricinfo.com/ci/content/stats/index.html
}
\references{
https://www.espncricinfo.com/ci/content/stats/index.html
}
\keyword{datasets}
