\name{maxwell}
\alias{maxwell}
\docType{data}
\title{
Data set for Glenn Maxwell
}
\description{
Data set for Glenn Maxwell
}
\usage{data("maxwell")}
\format{
  The format is:
 chr "maxwell"
}
\details{
Data set for Glenn Maxwell
}
\source{
https://www.espncricinfo.com/ci/content/stats/index.html
}
\references{
https://www.espncricinfo.com/ci/content/stats/index.html
}
\keyword{datasets}
