\name{mendis}
\alias{mendis}
\docType{data}
\title{
Data set for Ajantha Mendis
}
\description{
Data set for Ajantha Mendis
}
\usage{data("mendis")}
\format{
  The format is:
 chr "mendis"
}
\details{
Data set for Ajantha Mendis
}
\source{
https://www.espncricinfo.com/ci/content/stats/index.html
}
\references{
https://www.espncricinfo.com/ci/content/stats/index.html
}
\keyword{datasets}
