\name{mitchell}
\alias{mitchell}
\docType{data}
\title{
Data set for Mitchell Johnson
}
\description{
Data set for Mitchell Johnson
}
\usage{data("mitchell")}
\format{
  The format is:
 chr "mitchell"
}
\details{
Data set for Mitchell Johnson
}
\source{
http://www.espncricinfo.com/ci/content/stats/index.html
}
\references{
http://www.espncricinfo.com/ci/content/stats/index.html
}
\keyword{datasets}
