\name{murali}
\alias{murali}
\docType{data}
\title{
Data set for Muthiah Muralitharan
}
\description{
Data set for Muthiah Muralitharan
}
\usage{data("murali")}
\format{
  The format is:
 chr "murali"
}
\details{
Data set for Muthiah Muralitharan
}
\source{
https://www.espncricinfo.com/ci/content/stats/index.html
}
\references{
https://www.espncricinfo.com/ci/content/stats/index.html
}
\keyword{datasets}
