\name{narine}
\alias{narine}
\docType{data}
\title{
Data set for Sunil Narine
}
\description{
Data set for Sunil Narine
}
\usage{data("narine")}
\format{
  The format is:
 chr "narine"
}
\details{
Data set for Sunil Narine
}
\source{
http://www.espncricinfo.com/ci/content/stats/index.html
}
\references{
http://www.espncricinfo.com/ci/content/stats/index.html
}
\keyword{datasets}
