% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DL_funs.R
\name{objfun}
\alias{objfun}
\title{weighted least-squares}
\usage{
objfun(p, restab, weights = 1, ...)
}
\arguments{
\item{p}{Duckworth-Lewis parameters}

\item{restab}{resource table}

\item{weights}{}
}
\description{
weighted least-squares
}

