% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DL_funs.R
\name{optfun}
\alias{optfun}
\title{Find optimal D-L parameters}
\usage{
optfun(restab, weights, tot.runs, start = DL.par, control = list(maxit =
  1e+05, parscale = start), method = "BFGS")
}
\arguments{
\item{restab}{resource table}

\item{weights}{weights (typically counts)}

\item{tot.runs}{total runs for scaling}

\item{start}{starting parameters}

\item{control}{control list for \code{\link{optim}}}

\item{method}{optimization method}
}
\description{
Find optimal D-L parameters
}

