% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.restab}
\alias{plot.restab}
\title{plot resource tables}
\usage{
\method{plot}{restab}(x, type = c("heatmap", "levelplot", "gg_lines",
  "gg_tiles", "rgl"), xlab = "Overs remaining", ylab = "Wickets lost",
  line_mapping = NULL, factor_wicket = FALSE, ...)
}
\arguments{
\item{x}{resource table}

\item{type}{plot type}

\item{xlab}{}

\item{ylab}{}

\item{line_mapping}{mapping aesthetics for lines}

\item{dots}{(unused at present; for method consistency)}
}
\description{
plot resource tables
}

