\name{sehwag}
\alias{sehwag}
\docType{data}
\title{
Data set for Virendar Sehwag
}
\description{
Data set for Virendar Sehwag
}
\usage{data("sehwag")}
\format{
  The format is:
 chr "sehwag"
}
\details{
Data set for Virendar Sehwag
}
\source{
http://www.espncricinfo.com/ci/content/stats/index.html
}
\references{
http://www.espncricinfo.com/ci/content/stats/index.html
}
\keyword{datasets}
