\name{southee}
\alias{southee}
\docType{data}
\title{
Data set for Tim Southee
}
\description{
Data set for Tim Southee
}
\usage{data("southee")}
\format{
  The format is:
 chr "southee"
}
\details{
Data set for Tin Southee
}
\source{
http://www.espncricinfo.com/ci/content/stats/index.html
}
\references{
http://www.espncricinfo.com/ci/content/stats/index.html
}
\keyword{datasets}
