\name{steyn}
\alias{steyn}
\docType{data}
\title{
Data set for Dale Steyn
}
\description{
Data set for Dale Steyn
}
\usage{data("steyn")}
\format{
  The format is:
 chr "steyn"
}
\details{
Data set for Dale Steyn
}
\source{
http://www.espncricinfo.com/ci/content/stats/index.html
}
\references{
http://www.espncricinfo.com/ci/content/stats/index.html
}
\keyword{datasets}
