% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plots.R
\name{tab.to.df}
\alias{tab.to.df}
\title{rearrange resource table to data frame}
\usage{
tab.to.df(x)
}
\arguments{
\item{x}{resource table \emph{or} named list of resource tables}
}
\description{
rearrange resource table to data frame
}

