\name{tendulkar}
\alias{tendulkar}
\docType{data}
\title{
Data set for Sachin Tendulkar
}
\description{
Data set for Sachin Tendulkar
}
\usage{data("tendulkar")}
\format{
  The format is:
 chr "tendulkar"
}
\details{
Data set for Sachin Tendulkar
}
\source{
http://www.espncricinfo.com/ci/content/stats/index.html
}
\references{
http://www.espncricinfo.com/ci/content/stats/index.html
}
\keyword{datasets}
