\name{tendulkar2}
\alias{tendulkar2}
\docType{data}
\title{
Data set for Sachin Tendulkar
}
\description{
Data set for Sachin Tendulkar
}
\usage{data("tendulkar2")}
\format{
  The format is:
 chr "tendulkar2"
}
\details{
Data set for Sachin Tendulkar
}
\source{
https://www.espncricinfo.com/ci/content/stats/index.html
}
\references{
https://www.espncricinfo.com/ci/content/stats/index.html
}
\keyword{datasets}
