\name{tendulkarsp}
\alias{tendulkarsp}
\docType{data}
\title{
Data set for Sachin Tendulkar
}
\description{
Data set for Sachin Tendulkar
}
\usage{data("tendulkarsp")}
\format{
  The format is:
 chr "tendulkarsp"
}
\details{
Data set for Sachin Tendulkar
}
\source{
http://www.espncricinfo.com/ci/content/stats/index.html
}
\references{
http://www.espncricinfo.com/ci/content/stats/index.html
}
\keyword{datasets}
