\name{ashwin}
\alias{ashwin}
\docType{data}
\title{
Data set for Ravichandran Ashwin
}
\description{
CSV file Ravichandran Ashwin
}
\usage{data("ashwin")}
\format{
  The format is:
 chr "ashwin"
}
\details{
CSV file Ravichandran Ashwin
}
\source{
https://www.espncricinfo.com/ci/content/stats/index.html
}
\references{
https://www.espncricinfo.com/ci/content/stats/index.html
}
\keyword{datasets}
