\name{ashwin1}
\alias{ashwin1}
\docType{data}
\title{
Data set for Ravichander Ashwin
}
\description{
Data set for Ravichander Ashwin
}
\usage{data("ashwin1")}
\format{
  The format is:
 chr "ashwin1"
}
\details{
Data set for Ravichander Ashwin
}
\source{
http://www.espncricinfo.com/ci/content/stats/index.html
}
\references{
http://www.espncricinfo.com/ci/content/stats/index.html
}
\keyword{datasets}
