\name{cricketr-package}
\alias{cricketr-package}
\alias{cricketr}
\docType{package}
\title{
This package analyzes the performances of cricketers using ESPN Cricinfo Statsguru data. The analysis
can be done for Test, ODI and Twenty20 cricket for both batsman & bowlers
}
\description{
This package analyzes the performances of cricketers using ESPN Cricinfo Statsguru data. The analysis
can be done for Test, ODI and Twenty20 cricket for both batsman & bowlers
}
\details{
\tabular{ll}{
Package: \tab cricketr\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-07-03\cr
License: \tab MIT\cr
}
Details in my post https://gigadom.wordpress.com/2015/07/04/introducing-cricketr-a-r-package-to-analyze-performances-of-cricketers/
}
\author{
Tinniam V Ganesh

Maintainer: 
Tinniam V Ganesh
tvganesh.85@gmail.com
}

\keyword{ package }
\seealso{
https://www.youtube.com/edit?o=U&video_id=q9uMPFVsXsI
}

