\name{ganguly}
\alias{ganguly}
\docType{data}
\title{
Data set for Sourav Ganguly
}
\description{
Data set for Sourav Ganguly
}
\usage{data("ganguly")}
\format{
  The format is:
 chr "ganguly"
}
\details{
Data set for Sourav Ganguly
}
\source{
https://www.espncricinfo.com/ci/content/stats/index.html
}
\references{
https://www.espncricinfo.com/ci/content/stats/index.html
}
\keyword{datasets}
