\name{gayle}
\alias{gayle}
\docType{data}
\title{
Data set for Chris Gayle
}
\description{
Data set for Chris Gayle
}
\usage{data("gayle")}
\format{
  The format is:
 chr "gayle"
}
\details{
Data set for Chris Gayle
}
\source{
http://www.espncricinfo.com/ci/content/stats/index.html
}
\references{
http://www.espncricinfo.com/ci/content/stats/index.html
}
\keyword{datasets}
