\name{kohli}
\alias{kohli}
\docType{data}
\title{
Data set for Virat Kohli
}
\description{
CSV file Virat Kohl
}
\usage{data("kohli")}
\format{
  The format is:
 chr "kohli"
}
\details{
CSV file Virat Kohli
}
\source{
https://www.espncricinfo.com/ci/content/stats/index.html
}
\references{
https://www.espncricinfo.com/ci/content/stats/index.html
}
\keyword{datasets}
