\name{kumble}
\alias{kumble}
\docType{data}
\title{
Data set for Anil Kumble
}
\description{
Data set for Anil Kumble
}
\usage{data("kumble")}
\format{
  The format is:
 chr "kumble"
}
\details{
Data set for Anil Kumble
}
\source{
http://www.espncricinfo.com/ci/content/stats/index.html
}
\references{
http://www.espncricinfo.com/ci/content/stats/index.html
}
\keyword{datasets}
