\name{kumblesp}
\alias{kumblesp}
\docType{data}
\title{
Data set for Anil Kumble
}
\description{
Data set for Anil Kumble
}
\usage{data("kumblesp")}
\format{
  The format is:
 chr "kumblesp"
}
\details{
Data set for Anil Kumble
}
\source{
https://www.espncricinfo.com/ci/content/stats/index.html
}
\references{
https://www.espncricinfo.com/ci/content/stats/index.html
}
\keyword{datasets}
