\name{percentRuns}
\alias{percentRuns}
\title{
Calculate the percent runs in each run range
}
\description{
Calculate the percent runs in each 10 run range
}
\usage{
percentRuns(file)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{file}{
Input
}
}
\details{
More details can be found in my short video tutorial in Youtube
https://www.youtube.com/watch?v=q9uMPFVsXsI
}
\value{
None
}
\references{
http://www.espncricinfo.com/ci/content/stats/index.html\cr
https://gigadom.wordpress.com/
}
\author{
Tinniam V Ganesh
}
\note{
Maintainer: Tinniam V Ganesh <tvganesh.85@gmail.com>
}

