\name{warne}
\alias{warne}
\docType{data}
\title{
Data set for Shane Warne
}
\description{
Data set for Shane Warne
}
\usage{data("warne")}
\format{
  The format is:
 chr "warne"
}
\details{
Data set for Shane Warne
}
\source{
http://www.espncricinfo.com/ci/content/stats/index.html
}
\references{
http://www.espncricinfo.com/ci/content/stats/index.html
}
\keyword{datasets}
